Add more tests for client authentication methods
